I double checked and nav bar is still the same size as before.
Banner is no longer sticky and slides behind navbar.
Can I have specific words for the button?

 I had adjusted the height of the buttons because the 'interest form' button was bigger than the original button because of the words. Should the interest button go below the 'learn more about us' button, next to the button, or in the banner?

Is the spacing, padding, and font size correct for the rest of the page now? I was unable to see that on my end.

I believe the scroll behavior is now fixed.

Banner is now an <h2> tag and the words have been updated.